# api/openapi.yaml
openapi: 3.1.0
info:
  title: MaxPosty API
  version: 1.0.0
  description: |
    Canonical HTTP contract shared by the MaxPosty backend and frontend.
    Browser sessions are carried by the secure `maxstudio_session` cookie.
servers:
  - url: https://api.maxposty.ru/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /workspaces/{workspace_id}/posts/{post_id}/update-published:
    parameters:
      - {$ref: '#/components/parameters/WorkspaceId'}
      - {$ref: '#/components/parameters/PostId'}
    post:
      tags: [Posts]
      operationId: updateWorkspacePublishedPost
      summary: Replace the existing MAX publication with the approved current revision
      security: [{sessionCookie: []}]
      responses:
        '200': {$ref: '#/components/responses/Post'}
        default: {$ref: '#/components/responses/Problem'}
  /workspaces/{workspace_id}/posts/{post_id}/sync-max:
    parameters:
      - {$ref: '#/components/parameters/WorkspaceId'}
      - {$ref: '#/components/parameters/PostId'}
    post:
      tags: [Posts]
      operationId: syncWorkspaceMAXPublication
      summary: Refresh URL, views and pin state from MAX
      security: [{sessionCookie: []}]
      responses:
        '200': {$ref: '#/components/responses/Post'}
        default: {$ref: '#/components/responses/Problem'}
  /workspaces/{workspace_id}/posts/{post_id}/pin:
    parameters:
      - {$ref: '#/components/parameters/WorkspaceId'}
      - {$ref: '#/components/parameters/PostId'}
    post:
      tags: [Posts]
      operationId: pinWorkspacePost
      security: [{sessionCookie: []}]
      responses:
        '200': {$ref: '#/components/responses/Post'}
        default: {$ref: '#/components/responses/Problem'}
    delete:
      tags: [Posts]
      operationId: unpinWorkspacePost
      security: [{sessionCookie: []}]
      responses:
        '200': {$ref: '#/components/responses/Post'}
        default: {$ref: '#/components/responses/Problem'}
  /workspaces/{workspace_id}/posts/{post_id}/publication:
    parameters:
      - {$ref: '#/components/parameters/WorkspaceId'}
      - {$ref: '#/components/parameters/PostId'}
    delete:
      tags: [Posts]
      operationId: deleteWorkspacePublication
      summary: Delete the MAX message while preserving the local post and view history
      security: [{sessionCookie: []}]
      responses:
        '200': {$ref: '#/components/responses/Post'}
        default: {$ref: '#/components/responses/Problem'}
  /workspaces/{workspace_id}/posts/{post_id}/view-history:
    parameters:
      - {$ref: '#/components/parameters/WorkspaceId'}
      - {$ref: '#/components/parameters/PostId'}
    get:
      tags: [Posts]
      operationId: getWorkspacePostViewHistory
      security: [{sessionCookie: []}]
      parameters:
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 1000, default: 500}
        - name: before
          in: query
          schema: {type: string, format: date-time}
      responses:
        '200':
          description: Newest-first immutable MAX view observations
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/PostViewSnapshot'}
        default: {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: maxstudio_session
  parameters:
...
    WorkspaceId:
      name: workspace_id
      in: path
      required: true
      schema: {type: string, minLength: 1, maxLength: 128}
...
    PostId:
      name: post_id
      in: path
      required: true
      schema: {type: integer, format: int64, minimum: 1}
...
  responses:
    Problem:
      description: Request failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemEnvelope'
    Post:
      description: Updated post
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Post'
...
  schemas:
    ProblemEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/Problem'
    Problem:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details: {}
...
    PostReviewStatus:
      type: string
      enum: [draft, in_review, changes_requested, approved]
    PostViewSnapshot:
      type: object
      required: [id, post_id, max_message_id, views, captured_at]
      properties:
        id: {type: integer, format: int64}
        post_id: {type: integer, format: int64}
        max_message_id: {type: string}
        views: {type: integer, format: int64, minimum: 0}
        captured_at: {type: string, format: date-time}
...
    PostFormat:
      type: string
      enum: [markdown, html]
    PostStatus:
      type: string
      enum: [draft, scheduled, publishing, published, failed]
    PostLinkButton:
      type: object
      required: [text, url]
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 64
        url:
          type: string
          format: uri
    PostAttachment:
      type: object
      required:
        - id
        - type
        - position
        - url
        - processing_status
        - size_bytes
        - mime_type
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum: [image, video]
        position:
          type: integer
          minimum: 0
          maximum: 11
        url:
          type: string
        processing_status:
          type: string
          enum: [uploading, processing, ready, failed]
        size_bytes:
          type: integer
          format: int64
          minimum: 0
        mime_type:
          type: string
        width:
          type: [integer, 'null']
          minimum: 1
        height:
          type: [integer, 'null']
          minimum: 1
        duration_ms:
          type: [integer, 'null']
          format: int64
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Post:
      type: object
      required:
        - id
        - title
        - content
        - format
        - status
        - attachments
        - link_buttons
        - notify
        - disable_link_preview
        - max_is_pinned
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
        workspace_id:
          type: string
        title:
          type: string
        content:
          type: string
        format:
          $ref: '#/components/schemas/PostFormat'
        status:
          $ref: '#/components/schemas/PostStatus'
        channel_id:
          type: [integer, 'null']
          format: int64
        image_url:
          type: [string, 'null']
        image_prompt:
          type: [string, 'null']
        attachments:
          type: array
          maxItems: 12
          items:
            $ref: '#/components/schemas/PostAttachment'
        link_buttons:
          type: array
          items:
            $ref: '#/components/schemas/PostLinkButton'
        notify:
          type: boolean
        disable_link_preview:
          type: boolean
        scheduled_at:
          type: [string, 'null']
          format: date-time
        max_message_id:
          type: [string, 'null']
        max_message_url:
          type: [string, 'null']
          format: uri
        max_views:
          type: [integer, 'null']
          format: int64
          minimum: 0
        max_stats_synced_at:
          type: [string, 'null']
          format: date-time
        max_is_pinned:
          type: boolean
        last_error:
          type: [string, 'null']
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        published_at:
          type: [string, 'null']
          format: date-time
        review_status:
          $ref: '#/components/schemas/PostReviewStatus'
        current_revision_id:
          type: [integer, 'null']
          format: int64
...